Cleaning tests (#29)
* redoing path name in delete_files.py

* working on pathname for delete_files

* updated file_delete.py with correct path reference

* update file path reference in file_rename

* updated unittests for file_delete, file_rename

* added oneshot test for file_delete and file_rename

+++ chargemaster/data_cleaning/file_delete.py
@@ -3,34 +3,39 @@ It should be used before file_rename.py and a source directory should be added
 as path_name.
 """
 
-from os import renames, unlink
+#from os import renames, unlink
 import pathlib
+import os
 
 #This program works to delete files in scraper folder that are not data files.
 
-def delete_files(path_name):
+def delete_files():
     """ This script iterates through all hospital folders in the directory path_name
     and deletes files that are not .csv or .xlsx"""
-    path = pathlib.Path(path_name)
-    #print("This is the path: ", path)
-
+    path = pathlib.Path(__file__).parents[2]/"data"/"scraped_data"
+    isdir = os.path.isdir(path)
+    dirname1 = os.path.basename(path)
+    if isdir != True:
+        raise ValueError ("Not a valid directory")
+    if dirname1 != "scraped_data":
+        raise ValueError ("Wrong Directory Specified")
+    #print(path)
+    
     for folder in path.iterdir():
         if folder.is_dir():
-            #counter = 1
-            # folder_name = file.stem
-            # print("This is the folder name: ", folder_name)
+            
             for file in folder.iterdir():
                 if file.is_file():
                     if file.suffix == ".csv":
                        pass
-                       #print("this we keep: ", file.suffix)
+
                     elif file.suffix == ".xlsx":
                         pass
-                        #print("this we keep: ", file.suffix)
+
                     else:
-                        #print("this we delete: ", file.suffix)
+
                         file.unlink()
             
                     
 
-#delete_files(path_name)
+#delete_files()
